+++ explorer/src/lib/mock-data/gql-chain-info.d.ts
@@ -50,5 +50,6 @@ type GQLTransaction = {
     id: string;
     isDeploy: boolean;
     memo: string;
+    txType: string;
   };
 };