+++ explorer/CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement warning for stale market data [#1892]
 - Add tooltip to current and pending stake showing the exact amount [#2363]
 - Add `memo` and `isDeploy` fields in transactions [#2362]
+- Add `txType` fields in transactions and display it [#2347]
 
 ### Changed
 
@@ -65,6 +66,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#2348]: https://github.com/dusk-network/rusk/issues/2348
 [#2362]: https://github.com/dusk-network/rusk/issues/2362
 [#2363]: https://github.com/dusk-network/rusk/issues/2363
+[#2363]: https://github.com/dusk-network/rusk/issues/2347
 
 <!-- VERSIONS -->